third_party/WebKit/LayoutTests/fast/dom/attr-style-too-lazy-expected.txt

This test verifies that acquiring the 'style' attribute via a DOM Attr node still works correctly after the inline style has been modified through the CSSOM API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS styleAttr.value is 'background-color: white;'
PASS styleAttr.value is 'background-color: green;'
PASS oldStyleAttr.value is 'background-color: red;'
PASS successfullyParsed is true

TEST COMPLETE
